fix(workstation): match dragged tab icons and labels
