+++ docs/frontend-ui-audit-2026-09-07/WorkstationTabContent.md
@@ -1,0 +1,8 @@
+# WorkstationTabContent UI audit
+
+| Line                                                                            | Element                      | Verdict          | Reason                                                                                                                                                                  | Suggested change |
+| ------------------------------------------------------------------------------- | ---------------------------- | ---------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------- |
+| `src/modules/WorkStation/shared/TabBar/components/WorkstationTabContent.tsx:67` | Shared content renderer      | keep with reason | Reuses WorkstationTabIcon and TabLabelRowScrim; introduces no interactive control or duplicate sortable registration.                                                   | None.            |
+| `src/modules/WorkStation/shared/TabBar/components/WorkstationTabContent.tsx:84` | Typography and status colors | keep with reason | Preserves existing 13px labels, 11px status markers, and semantic Git colors. Available workstation typography presets add different font weights or serve other roles. | None.            |
+
+Verdict totals: **0 fix**, **2 keep with reason**, **0 abstract**.
