+++ src/modules/WorkStation/shared/TabBar/components/SortableTab/__tests__/index.test.ts
@@ -5,6 +5,7 @@ import { describe, expect, it, vi } from "vitest";
 import type { WorkStationTab } from "@src/store/workstation/tabs/types";
 
 import { SortableTab, resolveWorkstationTabIntegrationIcon } from "..";
+import { WorkstationTabContent } from "../../WorkstationTabContent";
 
 vi.mock("@src/components/IntegrationIcon", () => ({
   default: ({ type, size }: { type: string; size: number }) =>
@@ -87,5 +88,46 @@ describe("resolveWorkstationTabIntegrationIcon", () => {
     expect(markup).toMatch(
       /<svg[^>]*class="[^"]*text-text-1[^"]*"[^>]*data-icon="layout-grid"/
     );
+  });
+});
+
+describe("Workstation tab drag content", () => {
+  it("renders a tool glyph and localized label instead of the stored title", () => {
+    const markup = renderToStaticMarkup(
+      createElement(WorkstationTabContent, {
+        tab: {
+          ...tab("start"),
+          title: "Stale stored title",
+          icon: "LayoutGrid",
+        },
+        isActive: true,
+      })
+    );
+
+    expect(markup).toContain('data-icon="layout-grid"');
+    expect(markup).toContain("navigation:routes.launchpad");
+    expect(markup).not.toContain("Stale stored title");
+  });
+
+  it("preserves timeline revisions and the comparison glyph", () => {
+    const markup = renderToStaticMarkup(
+      createElement(WorkstationTabContent, {
+        tab: {
+          ...tab("git-diff", {
+            isTimeline: true,
+            shortSha: "abc123",
+            headShortSha: "def456",
+          }),
+          title: "index.ts",
+        },
+        isActive: false,
+      })
+    );
+
+    expect(markup).toContain("index.ts (abc123)");
+    expect(markup).toContain("(def456)");
+    expect(markup).toContain('data-icon="move-horizontal"');
+    expect(markup).toContain('data-icon="lock"');
+    expect(markup).not.toContain("Working Tree");
   });
 });
